Re-factoring - Use GoogleMaps below OL3
Having a GoogleMaps map purely working by writing OL3 code would
not work because of the different kinds of interactions required
to happen on the GMaps map.

This re-factoring introduces a new approach: do like what was made
in the past with ol3 v3.0.0, i.e. have GoogleMaps include a OL3
map as a control that takes 100% of the size.

This work is still in progress...