Fix empty style
It is valid to create an empty `ol.style.Style` object in ol3.
When parsing its properties, nothing is found (obviously, it's
empty), which resulted in giving an empty style options object
to Google Maps, which gets ignored. To fix that, the 'visible'
is set when the style options object is empty.