Fix \#4 - Deal with rotation
This patch introduces documentation about the limitations of this
library, notably the 'rotation' that should not be enabled, as
mentioned in \#4.

A new `LIMITATIONS.md` file was added to document this.

A new utility method that returns the default interactions that
include  only the ones that are recommended
(well, more like 'exclude' those that aren't).