Delay the activation of GoogleMaps
In some cases, for example when using a `ol.control.ZoomSlider`,
that control used to not work at all, because it re-renders the
map and use the size of the elemement to calculate size values.
While activating GoogleMaps, the ol3 map becomes in limbo, thus
caused the size to return `NaN`.

This patch makes sure that the ol3 map is ready before activating
Google Maps.