Dispatch event after setVisible on Google layer
This commit is a temporary fix for issue #60. It adds a setVisible
function to the Google layer class, which calls the original
parent function and dispatches a change:visible event. This is
because for an unknown reason, when both OLGM and OL3 are compiled
and used at the same time, the layer's observable properties don't
fire change events anymore for the Google layer.