Improve procedure for building examples
This commit adds a few improvements to the Makefile, specifically to the make dist-examples command. First, the script tags containing a link to the ol library are removed. They are no longer necessary as ol3 is compiled with ol3gm.

Second, the css file for openlayers located at `node_modules/openlayers/css/ol.css` is copied to the examples folder, and the  paths are updated in the HTML files.

The same thing is done for `css/ol3gm.css`

Finally, the api key can now be sent as a parameter when building examples. To do this, we assume the key already entered in the HTML document is the key for 127.0.0.1. We look for it and replace it with the variable defined in API_KEY, which has the key for 127.0.0.1 as its default value.

For example, output of `make API_KEY=AIzaSyCXIGZ41A54812NVcfTIuFS2mQUK-9JYeg dist-examples`:

```<script type="text/javascript"
    src="https://maps.googleapis.com/maps/api/js?v=3.21&key=AIzaSyCXIGZ41A54812NVcfTIuFS2mQUK-9JYeg"></script>```

Output for `make dist-examples`:

```<script type="text/javascript"
    src="https://maps.googleapis.com/maps/api/js?v=3.21&key=AIzaSyD71KlyTCXJouZsGbgPCJ-oCtK76fZJUTQ"></script>```