Fix invalid requests being sent on tile layers
What is going on around there:
`googleGetTileUrlFunction_` is the function called by Google Maps when it needs to load a tile for a set of coordinates. Our objective here is to return an url for the tile at the requested coordinates. We do that by "converting" the coordinates sent by Google's API to coordinates compatible with OL3. This step only implies a small operation on the y coordinate, and placing them in an array with the zoom level.

The problem occured because google maps checks for tiles at coordinates that shouldn't be valid, such as above the world map. We blindly build the url for these requests, even though they aren't valid.

The fix here is to check the extent for the tile at the coordinates requested, and check if that extent intersects with the layer's extent. If it does, we send the url as usual, otherwise we skip the tile.