Support minResolution and maxResolution
This fix adds support for layer attributes minResolution and maxResolution. Here's how it works:

On imageWMS and tile layer heralds, we already have a function to generate the tiles to render in Google Maps. With this fix, we check if we're within the mininum and maximum resolution before creating those tiles. If we're not, we simply return, so no requests are made.

On vector layers, we listen to the resolution change event (when we zoom in and out). Each time it happens, we check if we're within the minimum and maximum resolutions for that layer. We tell the result to the vectorFeatureHerald, which in turn gives the result to every feature it possess. When a feature is set invisible, we remove it from the Google Maps map, without deleting the feature itself. When we set it visible, we put it back on the back.