Support layer order for tile and image layers
This commit has the objective of ordering the layers the same way OpenLayer does.

Google Maps has several "panes" where it render things. Layers in one pane will always display above layers in the next one, no matter what. The first thing I had to do was to make sure the image and tile layers were in the same pane. The simpliest way to achieve that was to move the image layers to the mapPane pane, where tiled layers are rendered.

The next step was finding a way to make them display in the right order. To do that, we could either place them in the correct order in the mapPane, or set a zIndex on each of them. I chose the second solution.

For the image layers, it's pretty simple since we create the div. When we do, we set the zIndex if we can, otherwise we leave it at 0. There is a function to set the zIndex on the div later if we need to.

For tile layers, it's a little more complicated. Google Maps doesn't give us any way to access the divs where tiles are rendered. First, we need to access the pane element. The only way to do so is from objects inheriting google.maps.OverlayView. I created a PanesOverlay object, which doesn't do anything other than provide a way to access the mapPane. This object is instanciated in the tileSourceHerald. It can only access the pane once Google Maps is done loading, which means we can't set the zIndex on tile layers as we add them on the map, unless we delay their creation. I decided to not delay it and set the zIndex later instead. Once Google Maps is "idle", or whenever a layer is added or removed, we call the orderLayers function in the tileSourceHerald. In this function, we loop through every tile layer we're watching, and for each of them, we remove them and add them back in Google Maps. This provides us a way to guess which div they are rendered in. Then, we set a timeout to change the zIndex for that div, as Google Maps likes to set it back to 0 if we don't...

The example included is a good way to test it. The map rendered by OL3 and Google Maps should always be identical, other than the OSM and Google Maps layers.