Update Feature.js
Track updates to feature styles during the lifetime of a feature. Right now the olgm library only tracks the style at the creation time of a feature. Any update to a feature style is not reflected by the feature herald later on. The goal of this PR is to add this capability. Note that since OpenLayers does not explicitly send change events for styles, we have to rely on the generic change event at the feature level.