Build change (split from original PR branch that included all changes)
In a legacy ES5 environment without transpiling support, the current bundling of the OpenLayers library with OLGM seems to be missing some elements of OpenLayers (as if only actual parts of OpenLayers directly used by OLGM are included). In order to inject the missing parts of OL in the application, one needs to include it directly. This is creating issues since part of the OL runtime becomes effectively duplicated.

The goal of this PR is to have the OpenLayers library be treated as an external dependency such that no part of it is bundled in the olgm.js file prepared as the output from the build-legacy build path.

Some npm dependencies are also updated to later versions.