Bug fix
The current implementation of the View Herald tracks zoom changes by converting from the view resolution to an actual zoom level. This is done by doing a linear interpolation between the two closest non-fractional zoom levels. As long as non-fractional zoom levels are used this is OK since the interpolation is not really used. However for fractional zoom levels, the result of the interpolation does not correspond to the actual zoom level used as the source (in the OL map). This is causing features to appear at different places in the Google Map compared to the OpenLayers one since it does not use the same zoom level.

The fix is to use the same logarithm-based formula (as used in OpenLayers) to convert from a resolution to a zoom level. After this change the content of both maps align perfectly at fractional zoom levels.