Modify example to demonstrate use of pane
The concept examples has been modified to demonstrate how we can specify on which Google Maps pane we want a feature to reside. This is done by defining the usePane property in the feature with its value indicating the Google Maps pane name.

For example:

feature.usePane = 'overlayLayer';

By default, the markerLayer pane is used if not specified (this corresponds to the existing ol3-google-maps behavior).

Google Maps panes are in the following rendering order (top-most to bottom-most):

floatPane
overlayMouseTarget
markerLayer
overlayLayer
mapPane

To see the effect of specifying a pane for a feature, the three green 'hi' icons are now created to reside in the overlayLayer pane. Dragging them across the '42' icons will now show them fully beneath these icons while they would appear on top of the '42' blue circles if no pane was specified.